+++ comment/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import Comment
+
+admin.site.register(Comment)
