+++ question/models.py
@@ -15,20 +15,4 @@ class Question(models.Model):
     delete_status = models.BooleanField(default=False)
 
     def __str__(self):
-        return f"{self.title}, {self.body}"  # pragma: no cover
-
-
-class Comment(models.Model):
-    """This class represents the comment model."""
-
-    question = models.ForeignKey(Question, related_name='comments',
-                                 on_delete=models.CASCADE)
-    comment = models.TextField(unique=True)
-    created_by = models.ForeignKey(User, related_name='comments',
-                                   on_delete=models.CASCADE)
-    date_created = models.DateTimeField(auto_now_add=True)
-    date_modified = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        """Return a readable representation of the comment model instance."""
-        return (self.comment)
+        return "{}".format(self.title)
