+++ question/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from question.views import (Questions, OneQuestion, CommentList, CommentDetail)
+from question.views import Questions, OneQuestion
 
 urlpatterns = [
     path("<int:meetup_id>/questions/", Questions.as_view(), name="questions"),
@@ -9,8 +9,4 @@ urlpatterns = [
         OneQuestion.as_view(),
         name="question",
     ),
-    path("<int:meetup_id>/questions/<question_id>/comments",
-         CommentList.as_view(), name='comment'),
-    path("<int:meetup_id>/questions/<question_id>/comments/<int:pk>",
-         CommentDetail.as_view(), name='comment_detail')
 ]
