+++ questioner/settings.py
@@ -49,12 +49,14 @@ INSTALLED_APPS = [
     "rest_framework_swagger",
     "drf_yasg",
     "answer.apps.AnswersConfig",
+    "comment.apps.CommentConfig",
 ]
 
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': [
-        'rest_framework.authentication.SessionAuthentication',
-        "questioner.auth.auth_default.BearerAuthentication"
+        # 'rest_framework.authentication.SessionAuthentication',
+        "rest_framework.authentication.BasicAuthentication",
+        "questioner.auth.auth_default.BearerAuthentication",
     ],
     "EXCEPTION_HANDLER": "questioner.utils.exception_handler.custom_exception_handler",
 }
